Desktop/Casting Quality Prediction/requirements.txt
streamlit
numpy
tensorflow
scikit-image
opencv-python
Pillow
